fix pointer arithmetic

+++ demo/sum.c
@@ -1,7 +1,7 @@
-int sum(short *a, int n) {
-	int sum = 0;
+long sum(short *a, int n) {
+	long sum = 0;
 	int i = 0;
-	int *a_i;
+	short *a_i;
 	for (; i < n; i++) {
 		a_i = a + i;
 		sum += *a_i;
